+++ controllers/CategoriesController.php
@@ -32,6 +32,9 @@ class CategoriesController extends Controller
      */
     public function actionIndex()
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
         $dataProvider = new ActiveDataProvider([
             'query' => Categories::find(),
         ]);
@@ -48,6 +51,9 @@ class CategoriesController extends Controller
      */
     public function actionView($id)
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
         return $this->render('view', [
             'model' => $this->findModel($id),
         ]);
@@ -60,6 +66,9 @@ class CategoriesController extends Controller
      */
     public function actionCreate()
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
         $model = new Categories();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -79,6 +88,9 @@ class CategoriesController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -98,6 +110,9 @@ class CategoriesController extends Controller
      */
     public function actionDelete($id)
     {
+        if (\Yii::$app->user->isGuest) {
+            return $this->goHome();
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
